* feat: GET all Tenants endpoint

* refactor: update swagger

* feat: Delete buddies endpoint

* Bug: Unable to Publish Property (keeps on asking for profile to be completed)

* make property_id nullable

* fix: update dev environment

* fix: update pipeline

* fix: update rsync flag

* Update api/landlords endpoint (#21)

* Merge PR #23 - Create Endpoint to return the buddies of the landlord (#23)

* Merge PR (#24) - Assign a random property ID in portfolio import 

* Create Endpoint to return the buddies of the landlord

* Assign a random property ID in portfolio import

* Merges PR #26 - Create endpoint to return the buddies of the landlord (#26)

* Create Endpoint to return the buddies of the landlord

* Assign a random property ID in portfolio import

* Create and endpoint to delete buddies (single and multiple selection)

* Create and endpoint to delete buddies (single and multiple selection)

* Assign 8 digit random alphanumeric  property ID in portfolio import

* Merged PR #28 - Update api/landlords endpoint (#28)

* Merge PR #30 - Create an API endpoint to accept personal notes to a tenant/prospect (#30)

* fix: expired properties shown in published list

* Merge PR #32 - Create an API endpoint to accept personal notes to a tenant/prospect (#32)

* Merged PR #33 - Test case: Tenant view_ Invisible offer (#33)

* Merged PR #34 - Assign a Property ID automatically (#34)

* Merged PR #35 - Empty cards are not allowed (#35)

* Merged PR #36 - Crud notes

* Create an API endpoint to accept personal notes to a tenant/prospect

* tenant ID Validation added

tenant ID Validation added

* change firebase service account

* unread notification count

* change firebase service to correct one

* remove console logs

* Merged  PR #41 - Bug: translation tool (phrase.com) is disconnected with Breeze app

* Merged PR #42 - lang change serliazer and db (#42)

* Revert "Merged PR #42 - lang change serliazer and db (#42)"

This reverts commit 70fd576cb56f80acf8fae92c5493c13fb97eaef6.

* Merged PR #43 - schema change after phrase lang length chnage (#43)

* Merged PR #44 - Bug:Landlord-Published-Property Id can be duplicated

* Merged PR #45 - Create api endpoint to return the dashboard data (counts)   landlord 2

* Create API Endpoint to Return the Dashboard Data (Counts) - Landlord

* matches , decided count added

* Merged PR #46 - Summary_estates_condition_published

* Merged Summary estates condition published (#47)

* Summary_estates_condition_published

* Update MatchController.js - summary api

* tenant-new-design-endpoints

* filter estates by status

* Merged PR #51 - Dev 50 (#51)

* Summary_estates_condition_published

* Update MatchController.js - summary api

* estates get updated with filter

* PR Merged (#53) : estate api filters added

+++ app/Services/EstateService.js
@@ -65,9 +65,13 @@ class EstateService {
    *
    */
   static async createEstate(data, userId) {
+    const propertyId = data.property_id
+      ? data.property_id
+      : Math.random().toString(36).substr(2, 8).toUpperCase()
     return Estate.createItem({
       ...data,
       user_id: userId,
+      property_id: propertyId,
       status: STATUS_DRAFT,
     })
   }
@@ -76,7 +80,7 @@ class EstateService {
    *
    */
   static getEstates(params = {}) {
-    const query = Estate.query()
+    const query = Estate.query().withCount('visits').withCount('decided').withCount('invite')
     if (params.query) {
       query.where(function () {
         this.orWhere('estates.street', 'ilike', `%${params.query}%`)
@@ -86,6 +90,13 @@ class EstateService {
 
     if (params.status) {
       query.whereIn('estates.status', isArray(params.status) ? params.status : [params.status])
+    }
+
+    if(params.filter) {
+      query
+      .whereHas('matches', (query) => {
+            query.whereIn('status', params.filter)
+      })
     }
 
     return query.orderBy('estates.id', 'desc')
@@ -566,7 +577,12 @@ class EstateService {
    *
    */
   static async publishEstate(estate) {
-    await CompanyService.validateUserContacts(estate.user_id)
+    const User = use('App/Models/User')
+    const user = await User.query().where('id', estate.user_id).first()
+    if (!user) return
+    if (user.company_id != null) {
+      await CompanyService.validateUserContacts(estate.user_id)
+    }
     await props({
       delMatches: () => Database.table('matches').where({ estate_id: estate.id }).delete(),
       delLikes: () => Database.table('likes').where({ estate_id: estate.id }).delete(),
